+++ include/AvatarOptimizer.h
@@ -14,19 +14,19 @@ namespace ark {
         AvatarOptimizer(Avatar& ava, const CameraIntrin& intrin, const cv::Size& image_size);
 
         /** Begin full optimization on the target data cloud */
-        void optimize(const Eigen::Matrix<double, 3, Eigen::Dynamic>& data_cloud, int icp_iters = 1, int num_threads = 4);
+        void optimize(const Eigen::Matrix<float, 3, Eigen::Dynamic>& data_cloud, int icp_iters = 1, int num_threads = 4);
 
         /** Begin align optimization to fit joints to target joint positions */
-        void align(const Eigen::Matrix<double, 3, Eigen::Dynamic>& smpl_joints, int icp_iters = 1, int num_threads = 4);
+        void align(const Eigen::Matrix<float, 3, Eigen::Dynamic>& smpl_joints, int icp_iters = 1, int num_threads = 4);
 
         /** Rotation representation size */
         static const int ROT_SIZE = 4;
 
         /** Optimization parameter r */
-        std::vector<Eigen::Quaterniond, Eigen::aligned_allocator<Eigen::Quaterniond> > r;
+        std::vector<Eigen::Quaternionf, Eigen::aligned_allocator<Eigen::Quaternionf> > r;
 
         /** Cost function component weights */
-        double betaPose = 0.1, betaShape = 1.0, betaJoints = 0.2;
+        float betaPose = 0.1f, betaShape = 1.0f, betaJoints = 0.2f;
 
         /** The avatar we are optimizing */
         Avatar& ava;
